+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/main.js
@@ -153,7 +153,7 @@ define([
                 }.bind(this));
         },
 
-        loadData: function() {
+        loadData: function(template) {
             var promise = $.Deferred();
             if (!this.content) {
                 this.content = new Content({id: this.options.id});
@@ -164,6 +164,7 @@ define([
                     this.options.webspace,
                     this.options.language,
                     true,
+                    template,
                     {
                         success: function(content) {
                             this.data = content.toJSON();
@@ -195,7 +196,17 @@ define([
             }, this);
 
             // getter for content data
-            this.sandbox.on('sulu.content.contents.get-data', function(callback) {
+            this.sandbox.on('sulu.content.contents.get-data', function(callback, reloadData, template) {
+                if (reloadData === true && template) {
+                    this.loadData(template).then(function() {
+                        // deep copy of object
+                        callback(this.sandbox.util.deepCopy(this.data), this.preview);
+
+                    }.bind(this));
+
+                    return;
+                }
+
                 // deep copy of object
                 callback(this.sandbox.util.deepCopy(this.data), this.preview);
             }.bind(this));