+++ src/Sulu/Bundle/MediaBundle/Api/Media.php
@@ -406,16 +406,21 @@ class Media extends ApiWrapper
     public function getVersions()
     {
         $versions = [];
+
+        $file = $this->getFile();
+
         /** @var FileVersion $fileVersion */
-        foreach ($this->getFile()->getFileVersions() as $fileVersion) {
+        foreach ($file->getFileVersions() as $fileVersion) {
             $versionData = [];
             if (isset($this->additionalVersionData[$fileVersion->getVersion()])) {
                 $versionData = $this->additionalVersionData[$fileVersion->getVersion()];
             }
+
             $versionData['version'] = $fileVersion->getVersion();
             $versionData['name'] = $fileVersion->getName();
             $versionData['created'] = $fileVersion->getCreated();
             $versionData['changed'] = $fileVersion->getChanged();
+            $versionData['active'] = $fileVersion->isActive();
             $versions[$fileVersion->getVersion()] = $versionData;
         }
 
@@ -865,13 +870,11 @@ class Media extends ApiWrapper
             } else {
                 $version = $file->getVersion();
             }
-            /** @var FileVersion $fileVersion */
-            foreach ($file->getFileVersions() as $fileVersion) {
-                if ($fileVersion->getVersion() == $version) {
-                    $this->fileVersion = $fileVersion;
 
-                    return $fileVersion;
-                }
+            if ($fileVersion = $file->getFileVersion($version)) {
+                $this->fileVersion = $fileVersion;
+
+                return $fileVersion;
             }
             break; // currently only one file per media exists
         }
@@ -898,7 +901,7 @@ class Media extends ApiWrapper
             return $this->file;
         }
 
-        throw new FileNotFoundException($this->entity->getId(), $this->version);
+        throw new FileNotFoundException($this->entity->getId());
     }
 
     /**